Ex4 add doc; use new conda env
This update to Example 4 adds additional markdown, code comments, and a few extra displays of outputs to better serve as a tutorial or learning guide. The code is essentially unchanged. The markdown also provides guidance on how to run this from a complete, custom, conda environement that includes all dependencies.