Restructuring
Moves the PhreeqcBMI and Mf6API classes into separate files. This is necessary to pave the way to a factor pattern implementation for grid transformation, which will depended on these classes. Defining them in the mf6rtm file, would lead to a circular reference.

This also simplifies the namespace and the mf6trm file.